backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
pydantic-settings==2.0.3
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
sqlalchemy==2.0.23
pyyaml==6.0.1
jsonschema==4.19.1
bcrypt==4.0.1
email-validator==2.1.0.post1
python-dotenv==1.0.0
